refactor: harden auth boundaries and split api helpers
