+++ backend-spring/src/main/java/com/myway/backendspring/api/AdminAssignmentsController.java
@@ -2,12 +2,13 @@ package com.myway.backendspring.api;
 
 import com.myway.backendspring.auth.SessionService;
 import com.myway.backendspring.auth.SessionView;
+import com.myway.backendspring.auth.RolePolicy;
+import com.myway.backendspring.api.support.ApiAuthGuards;
 import com.myway.backendspring.common.ApiResponse;
 import com.myway.backendspring.feature.FeatureStoreService;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -28,31 +29,17 @@ public class AdminAssignmentsController {
         this.featureStore = featureStore;
     }
 
-    private SessionView require(String auth) {
-        return sessionService.me(auth);
-    }
-    private <T> ResponseEntity<ApiResponse<T>> unauthenticated() {
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(ApiResponse.failure("UNAUTHENTICATED", "로그인이 필요합니다."));
-    }
-    private <T> ResponseEntity<ApiResponse<T>> forbidden() {
-        return ResponseEntity.status(HttpStatus.FORBIDDEN).body(ApiResponse.failure("FORBIDDEN", "관리자 권한이 필요합니다."));
-    }
-
-    private boolean isAdmin(SessionView session) {
-        return session != null && "admin".equals(session.user().role());
-    }
-
     @GetMapping("/{courseId}")
     public ResponseEntity<ApiResponse<Map<String, Object>>> getAssignment(
             @RequestHeader(value = "Authorization", required = false) String auth,
             @PathVariable String courseId
     ) {
-        SessionView session = require(auth);
+        SessionView session = ApiAuthGuards.requireSession(sessionService, auth);
         if (session == null) {
-            return unauthenticated();
+            return ApiAuthGuards.unauthenticated();
         }
-        if (!isAdmin(session)) {
-            return forbidden();
+        if (!RolePolicy.isAdmin(session.user().role())) {
+            return ApiAuthGuards.forbidden("관리자 권한이 필요합니다.");
         }
         return ResponseEntity.ok(ApiResponse.success(featureStore.getAdminAssignment(courseId)));
     }
@@ -63,12 +50,12 @@ public class AdminAssignmentsController {
             @PathVariable String courseId,
             @Valid @RequestBody AssignmentUpdateRequest body
     ) {
-        SessionView session = require(auth);
+        SessionView session = ApiAuthGuards.requireSession(sessionService, auth);
         if (session == null) {
-            return unauthenticated();
+            return ApiAuthGuards.unauthenticated();
         }
-        if (!isAdmin(session)) {
-            return forbidden();
+        if (!RolePolicy.isAdmin(session.user().role())) {
+            return ApiAuthGuards.forbidden("관리자 권한이 필요합니다.");
         }
 
         List<String> studentIds = normalizeStudentIds(body.student_ids());
