+++ backend-spring/src/main/java/com/myway/backendspring/api/AiController.java
@@ -1,6 +1,7 @@
 package com.myway.backendspring.api;
 
 import com.fasterxml.jackson.annotation.JsonAnySetter;
+import com.myway.backendspring.api.support.AiRequestSupport;
 import com.myway.backendspring.auth.SessionService;
 import com.myway.backendspring.auth.SessionView;
 import com.myway.backendspring.common.ApiResponse;
@@ -16,7 +17,6 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.ArrayList;
 import java.util.stream.Collectors;
 
 @RestController
@@ -30,7 +30,6 @@ public class AiController {
     public record AnswerRequest(@NotBlank String question, String lecture_id) {}
     public record SummaryRequest(@NotBlank String lecture_id, String style, String language) {}
     public record QuizRequest(@NotBlank String lecture_id) {}
-    private record RagScope(String lectureId, String courseId) {}
     public static final class AiSettingsUpdateRequest {
         private Map<String, Object> settings;
         private final Map<String, Object> extras = new HashMap<>();
@@ -80,12 +79,20 @@ public class AiController {
     private final FeatureStoreService featureStore;
     private final DemoLearningService learningService;
     private final AiRuntimeService aiRuntimeService;
+    private final AiRequestSupport aiRequestSupport;
 
-    public AiController(SessionService sessionService, FeatureStoreService featureStore, DemoLearningService learningService, AiRuntimeService aiRuntimeService) {
+    public AiController(
+            SessionService sessionService,
+            FeatureStoreService featureStore,
+            DemoLearningService learningService,
+            AiRuntimeService aiRuntimeService,
+            AiRequestSupport aiRequestSupport
+    ) {
         this.sessionService = sessionService;
         this.featureStore = featureStore;
         this.learningService = learningService;
         this.aiRuntimeService = aiRuntimeService;
+        this.aiRequestSupport = aiRequestSupport;
     }
 
     private SessionView require(String auth) { return sessionService.me(auth); }
@@ -151,10 +158,10 @@ public class AiController {
         if (session == null) return unauthenticated();
         if (!featureStore.canConsumeAi(session.user().id())) return dailyLimitExceeded();
 
-        String query = normalize(body.query());
+        String query = aiRequestSupport.normalize(body.query());
 
-        RagScope scope = resolveRagScope(body.lecture_id(), body.course_id());
-        ResponseEntity<ApiResponse<Map<String, Object>>> scopeError = validateRagScope(scope);
+        AiRequestSupport.RagScope scope = aiRequestSupport.resolveRagScope(body.lecture_id(), body.course_id());
+        ResponseEntity<ApiResponse<Map<String, Object>>> scopeError = aiRequestSupport.validateRagScope(scope);
         if (scopeError != null) return scopeError;
 
         Integer limit = body.limit();
@@ -162,8 +169,8 @@ public class AiController {
         boolean includeDebug = Boolean.TRUE.equals(body.include_debug());
         Map<String, Object> data = featureStore.ragOverview(
                 query,
-                optionalNormalized(scope.lectureId()),
-                optionalNormalized(scope.courseId()),
+                aiRequestSupport.optionalNormalized(scope.lectureId()),
+                aiRequestSupport.optionalNormalized(scope.courseId()),
                 limit,
                 minScore,
                 includeDebug
@@ -181,8 +188,8 @@ public class AiController {
         SessionView session = require(auth);
         if (session == null) return unauthenticated();
         Map<String, Object> data = featureStore.ragIndexOverview(
-                optionalNormalized(lectureId),
-                optionalNormalized(courseId)
+                aiRequestSupport.optionalNormalized(lectureId),
+                aiRequestSupport.optionalNormalized(courseId)
         );
         return ResponseEntity.ok(ApiResponse.success(data, "RAG 인덱스 현황을 조회했습니다."));
     }
@@ -194,12 +201,12 @@ public class AiController {
     ) {
         SessionView session = require(auth);
         if (session == null) return unauthenticated();
-        RagScope scope = resolveRagScope(body.lecture_id(), body.course_id());
-        ResponseEntity<ApiResponse<Map<String, Object>>> scopeError = validateRagScope(scope);
+        AiRequestSupport.RagScope scope = aiRequestSupport.resolveRagScope(body.lecture_id(), body.course_id());
+        ResponseEntity<ApiResponse<Map<String, Object>>> scopeError = aiRequestSupport.validateRagScope(scope);
         if (scopeError != null) return scopeError;
         Map<String, Object> data = featureStore.rebuildRagIndex(
-                optionalNormalized(scope.lectureId()),
-                optionalNormalized(scope.courseId())
+                aiRequestSupport.optionalNormalized(scope.lectureId()),
+                aiRequestSupport.optionalNormalized(scope.courseId())
         );
         return ResponseEntity.status(HttpStatus.CREATED).body(ApiResponse.success(data, "RAG 인덱스를 재생성했습니다."));
     }
@@ -211,12 +218,12 @@ public class AiController {
     ) {
         SessionView session = require(auth);
         if (session == null) return unauthenticated();
-        RagScope scope = resolveRagScope(body.lecture_id(), body.course_id());
-        ResponseEntity<ApiResponse<Map<String, Object>>> scopeError = validateRagScope(scope);
+        AiRequestSupport.RagScope scope = aiRequestSupport.resolveRagScope(body.lecture_id(), body.course_id());
+        ResponseEntity<ApiResponse<Map<String, Object>>> scopeError = aiRequestSupport.validateRagScope(scope);
         if (scopeError != null) return scopeError;
         Map<String, Object> data = featureStore.clearRagIndex(
-                optionalNormalized(scope.lectureId()),
-                optionalNormalized(scope.courseId())
+                aiRequestSupport.optionalNormalized(scope.lectureId()),
+                aiRequestSupport.optionalNormalized(scope.courseId())
         );
         return ResponseEntity.ok(ApiResponse.success(data, "RAG 인덱스를 초기화했습니다."));
     }
@@ -241,7 +248,7 @@ public class AiController {
         SessionView session = require(auth);
         if (session == null) return unauthenticated();
         if (!featureStore.canConsumeAi(session.user().id())) return dailyLimitExceeded();
-        String message = normalize(body.message());
+        String message = aiRequestSupport.normalize(body.message());
         Map<String, Object> runtime = aiRuntimeService.generate(
                 "intent",
                 "메시지 의도를 한 단어로 분류하고 이유를 한 줄로 설명하세요: " + message,
@@ -252,7 +259,7 @@ public class AiController {
         data.put("confidence", 0.82);
         data.put("action", "show_recommendations");
         data.put("reason", runtime.getOrDefault("text", "Spring demo intent classifier"));
-        data.put("lecture_id", optionalNormalized(body.lecture_id()));
+        data.put("lecture_id", aiRequestSupport.optionalNormalized(body.lecture_id()));
         data.put("provider", runtime.getOrDefault("provider", "demo"));
         data.put("model", runtime.getOrDefault("model", "demo-intent-v1"));
         data.put("live", runtime.getOrDefault("live", false));
@@ -266,9 +273,9 @@ public class AiController {
         SessionView session = require(auth);
         if (session == null) return unauthenticated();
         if (!featureStore.canConsumeAi(session.user().id())) return dailyLimitExceeded();
-        String query = normalize(body.query());
-        String lectureId = optionalNormalized(body.lecture_id());
-        ResponseEntity<ApiResponse<Map<String, Object>>> lectureError = validateLecture(body.lecture_id());
+        String query = aiRequestSupport.normalize(body.query());
+        String lectureId = aiRequestSupport.optionalNormalized(body.lecture_id());
+        ResponseEntity<ApiResponse<Map<String, Object>>> lectureError = aiRequestSupport.validateLecture(body.lecture_id());
         if (lectureError != null) return lectureError;
         Map<String, Object> runtime = aiRuntimeService.generate(
                 "search",
@@ -278,7 +285,7 @@ public class AiController {
         Map<String, Object> data = new HashMap<>();
         data.put("query", query);
         data.put("hits", List.of(Map.of("id", "hit-1", "title", String.valueOf(runtime.getOrDefault("text", "Spring Boot 시작")), "score", 0.91)));
-        data.put("sources", resolveRagSources(query, lectureId, null, 4));
+        data.put("sources", aiRequestSupport.resolveRagSources(query, lectureId, null, 4));
         data.put("provider", runtime.getOrDefault("provider", "demo"));
         data.put("model", runtime.getOrDefault("model", "demo-search-v1"));
         data.put("live", runtime.getOrDefault("live", false));
@@ -292,12 +299,12 @@ public class AiController {
         SessionView session = require(auth);
         if (session == null) return unauthenticated();
         if (!featureStore.canConsumeAi(session.user().id())) return dailyLimitExceeded();
-        String question = normalize(body.question());
-        String lectureId = optionalNormalized(body.lecture_id());
-        ResponseEntity<ApiResponse<Map<String, Object>>> lectureError = validateLecture(body.lecture_id());
+        String question = aiRequestSupport.normalize(body.question());
+        String lectureId = aiRequestSupport.optionalNormalized(body.lecture_id());
+        ResponseEntity<ApiResponse<Map<String, Object>>> lectureError = aiRequestSupport.validateLecture(body.lecture_id());
         if (lectureError != null) return lectureError;
         Map<String, Object> runtime = aiRuntimeService.generate("answer", question, featureStore.aiSettings(session.user().id()));
-        List<Map<String, Object>> sources = resolveRagSources(question, lectureId, null, 4);
+        List<Map<String, Object>> sources = aiRequestSupport.resolveRagSources(question, lectureId, null, 4);
         Map<String, Object> data = new HashMap<>();
         data.put("answer", runtime.getOrDefault("text", "[Spring AI 응답] " + question));
         data.put("sources", sources);
@@ -321,10 +328,10 @@ public class AiController {
         SessionView session = require(auth);
         if (session == null) return unauthenticated();
         if (!featureStore.canConsumeAi(session.user().id())) return dailyLimitExceeded();
-        String lectureId = requireLectureId(body.lecture_id());
+        String lectureId = aiRequestSupport.requireLectureId(body.lecture_id());
         if (learningService.getLecture(lectureId) == null) return ResponseEntity.status(HttpStatus.NOT_FOUND).body(ApiResponse.failure("LECTURE_NOT_FOUND", "강의를 찾을 수 없습니다."));
-        String style = defaultIfBlank(body.style(), "brief");
-        String language = defaultIfBlank(body.language(), "ko");
+        String style = aiRequestSupport.defaultIfBlank(body.style(), "brief");
+        String language = aiRequestSupport.defaultIfBlank(body.language(), "ko");
         Map<String, Object> runtime = aiRuntimeService.generate(
                 "summary",
                 lectureId + " 강의 내용을 " + style + " 스타일로 " + language + " 요약",
@@ -348,7 +355,7 @@ public class AiController {
         SessionView session = require(auth);
         if (session == null) return unauthenticated();
         if (!featureStore.canConsumeAi(session.user().id())) return dailyLimitExceeded();
-        String lectureId = requireLectureId(body.lecture_id());
+        String lectureId = aiRequestSupport.requireLectureId(body.lecture_id());
         if (learningService.getLecture(lectureId) == null) return ResponseEntity.status(HttpStatus.NOT_FOUND).body(ApiResponse.failure("LECTURE_NOT_FOUND", "강의를 찾을 수 없습니다."));
         Map<String, Object> runtime = aiRuntimeService.generate(
                 "quiz",
@@ -366,142 +373,4 @@ public class AiController {
         return ResponseEntity.ok(ApiResponse.success(data, "퀴즈가 생성되었습니다."));
     }
 
-    private ResponseEntity<ApiResponse<Map<String, Object>>> validateLecture(String lectureIdRaw) {
-        String lectureId = lectureIdRaw == null ? "" : lectureIdRaw.trim();
-        if (!lectureId.isBlank() && learningService.getLecture(lectureId) == null) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(ApiResponse.failure("LECTURE_NOT_FOUND", "강의를 찾을 수 없습니다."));
-        }
-        return null;
-    }
-
-    private RagScope resolveRagScope(String lectureIdRaw, String courseIdRaw) {
-        return new RagScope(normalize(lectureIdRaw), normalize(courseIdRaw));
-    }
-
-    private ResponseEntity<ApiResponse<Map<String, Object>>> validateRagScope(RagScope scope) {
-        if (scope.lectureId().isBlank() && scope.courseId().isBlank()) {
-            return ResponseEntity.badRequest().body(ApiResponse.failure("LECTURE_OR_COURSE_REQUIRED", "lecture_id 또는 course_id가 필요합니다."));
-        }
-        if (!scope.lectureId().isBlank() && learningService.getLecture(scope.lectureId()) == null) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(ApiResponse.failure("LECTURE_NOT_FOUND", "강의를 찾을 수 없습니다."));
-        }
-        if (!scope.courseId().isBlank() && learningService.getCourseLectures(scope.courseId()).isEmpty()) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(ApiResponse.failure("COURSE_NOT_FOUND", "강의를 찾을 수 없습니다."));
-        }
-        return null;
-    }
-
-    private String normalize(String value) {
-        return value == null ? "" : value.trim();
-    }
-
-    private String optionalNormalized(String value) {
-        String normalized = normalize(value);
-        return normalized.isBlank() ? null : normalized;
-    }
-
-    private String defaultIfBlank(String value, String fallback) {
-        String normalized = normalize(value);
-        return normalized.isBlank() ? fallback : normalized;
-    }
-
-    private String requireLectureId(String lectureIdRaw) {
-        String lectureId = normalize(lectureIdRaw);
-        return lectureId.isBlank() ? null : lectureId;
-    }
-
-    private List<Map<String, Object>> resolveRagSources(String query, String lectureId, String courseId, int limit) {
-        try {
-            Map<String, Object> rag = featureStore.ragOverview(query, lectureId, courseId, limit, 0.0, false);
-            List<Map<String, Object>> chunks = extractChunkList(rag);
-            List<Map<String, Object>> sources = new ArrayList<>();
-            for (Map<String, Object> chunk : chunks) {
-                String sourceLectureId = normalize(String.valueOf(chunk.getOrDefault("lecture_id", lectureId == null ? "" : lectureId)));
-                int startMs = asInt(chunk.get("start_ms"));
-                int endMs = asInt(chunk.get("end_ms"));
-                String text = firstNotBlank(
-                        String.valueOf(chunk.getOrDefault("excerpt", "")),
-                        String.valueOf(chunk.getOrDefault("content", "")),
-                        String.valueOf(chunk.getOrDefault("title", ""))
-                );
-                double score = asDouble(chunk.get("similarity"));
-                Map<String, Object> source = new HashMap<>();
-                source.put("lecture_id", sourceLectureId.isBlank() ? lectureId : sourceLectureId);
-                source.put("start_ms", Math.max(0, startMs));
-                source.put("end_ms", Math.max(Math.max(0, startMs), endMs));
-                source.put("text", text);
-                source.put("score", score);
-                sources.add(source);
-            }
-            if (!sources.isEmpty()) {
-                return sources;
-            }
-        } catch (Exception ignored) {
-            // Keep response contract stable even when RAG retrieval fails.
-        }
-        return List.of(Map.of(
-                "lecture_id", lectureId == null ? "" : lectureId,
-                "start_ms", 0,
-                "end_ms", 0,
-                "text", "근거를 준비 중입니다.",
-                "score", 0.0
-        ));
-    }
-
-    @SuppressWarnings("unchecked")
-    private List<Map<String, Object>> extractChunkList(Map<String, Object> ragPayload) {
-        if (ragPayload == null || ragPayload.isEmpty()) {
-            return List.of();
-        }
-        Object chunks = ragPayload.get("chunks");
-        if (chunks instanceof List<?> list) {
-            return list.stream()
-                    .filter(Map.class::isInstance)
-                    .map(item -> (Map<String, Object>) item)
-                    .toList();
-        }
-        Object search = ragPayload.get("search");
-        if (search instanceof Map<?, ?> searchMap) {
-            Object hits = searchMap.get("hits");
-            if (hits instanceof List<?> list) {
-                return list.stream()
-                        .filter(Map.class::isInstance)
-                        .map(item -> (Map<String, Object>) item)
-                        .toList();
-            }
-        }
-        return List.of();
-    }
-
-    private int asInt(Object value) {
-        if (value instanceof Number number) {
-            return number.intValue();
-        }
-        try {
-            return Integer.parseInt(String.valueOf(value));
-        } catch (Exception ignored) {
-            return 0;
-        }
-    }
-
-    private double asDouble(Object value) {
-        if (value instanceof Number number) {
-            return number.doubleValue();
-        }
-        try {
-            return Double.parseDouble(String.valueOf(value));
-        } catch (Exception ignored) {
-            return 0.0;
-        }
-    }
-
-    private String firstNotBlank(String... values) {
-        for (String value : values) {
-            String normalized = normalize(value);
-            if (!normalized.isBlank()) {
-                return normalized;
-            }
-        }
-        return "";
-    }
 }
