+++ backend-spring/src/main/java/com/myway/backendspring/api/EnrollmentsController.java
@@ -2,6 +2,7 @@ package com.myway.backendspring.api;
 
 import com.myway.backendspring.auth.SessionService;
 import com.myway.backendspring.auth.SessionView;
+import com.myway.backendspring.api.support.ApiAuthGuards;
 import com.myway.backendspring.common.ApiResponse;
 import com.myway.backendspring.domain.*;
 import jakarta.validation.Valid;
@@ -25,27 +26,20 @@ public class EnrollmentsController {
     }
 
     public record EnrollmentRequest(@NotBlank String courseId) {}
-    private SessionView require(String auth) {
-        return sessionService.me(auth);
-    }
-    private <T> ResponseEntity<ApiResponse<T>> unauthenticated() {
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(ApiResponse.failure("UNAUTHENTICATED", "로그인이 필요합니다."));
-    }
-
     @GetMapping
     public ResponseEntity<ApiResponse<List<EnrollmentItem>>> list(@RequestHeader(value = "Authorization", required = false) String auth) {
-        SessionView session = require(auth);
+        SessionView session = ApiAuthGuards.requireSession(sessionService, auth);
         if (session == null) {
-            return unauthenticated();
+            return ApiAuthGuards.unauthenticated();
         }
         return ResponseEntity.ok(ApiResponse.success(learningService.listEnrollments(session.user().id())));
     }
 
     @PostMapping
     public ResponseEntity<ApiResponse<Map<String, Object>>> enroll(@RequestHeader(value = "Authorization", required = false) String auth, @Valid @RequestBody EnrollmentRequest body) {
-        SessionView session = require(auth);
+        SessionView session = ApiAuthGuards.requireSession(sessionService, auth);
         if (session == null) {
-            return unauthenticated();
+            return ApiAuthGuards.unauthenticated();
         }
 
         CourseDetail course = learningService.getCourseDetail(body.courseId().trim(), session.user().id());
