+++ backend-spring/src/main/java/com/myway/backendspring/api/LecturesController.java
@@ -2,6 +2,7 @@ package com.myway.backendspring.api;
 
 import com.myway.backendspring.auth.SessionService;
 import com.myway.backendspring.auth.SessionView;
+import com.myway.backendspring.api.support.ApiAuthGuards;
 import com.myway.backendspring.common.ApiResponse;
 import com.myway.backendspring.domain.DemoLearningService;
 import com.myway.backendspring.domain.LectureItem;
@@ -33,9 +34,9 @@ public class LecturesController {
 
     @PostMapping("/{lectureId}/complete")
     public ResponseEntity<ApiResponse<Map<String, Object>>> complete(@PathVariable String lectureId, @RequestHeader(value = "Authorization", required = false) String auth) {
-        SessionView session = require(auth);
+        SessionView session = ApiAuthGuards.requireSession(sessionService, auth);
         if (session == null) {
-            return unauthenticated();
+            return ApiAuthGuards.unauthenticated();
         }
 
         Map<String, Object> result = learningService.completeLecture(session.user().id(), lectureId);
@@ -47,13 +48,5 @@ public class LecturesController {
         }
 
         return ResponseEntity.ok(ApiResponse.success(result, "강의 진도가 저장되었습니다."));
-    }
-
-    private SessionView require(String auth) {
-        return sessionService.me(auth);
-    }
-
-    private ResponseEntity<ApiResponse<Map<String, Object>>> unauthenticated() {
-        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(ApiResponse.failure("UNAUTHENTICATED", "로그인이 필요합니다."));
     }
 }
