+++ frontend/src/lib/api-admin-assignments.ts
@@ -1,4 +1,4 @@
-import { request, unwrap } from './api-core';
+import { getStoredAuth, request } from './api-core';
 
 export type AdminAssignmentRecord = {
   course_id: string;
@@ -7,26 +7,23 @@ export type AdminAssignmentRecord = {
   updated_at?: string;
 };
 
-function fallbackRecord(courseId: string): AdminAssignmentRecord {
-  return {
-    course_id: courseId,
-    student_ids: [],
-  };
-}
-
-export async function loadAdminAssignment(courseId: string): Promise<AdminAssignmentRecord> {
+export async function loadAdminAssignment(courseId: string): Promise<AdminAssignmentRecord | null> {
+  const token = getStoredAuth()?.session_token ?? null;
+  if (!token) return null;
   const response = await request<AdminAssignmentRecord>(`/api/v1/admin/assignments/${courseId}`, {
     method: 'GET',
-  });
-  return unwrap(response, () => fallbackRecord(courseId));
+  }, token);
+  return response?.success && response.data ? response.data : null;
 }
 
-export async function saveAdminAssignment(courseId: string, studentIds: string[]): Promise<AdminAssignmentRecord> {
+export async function saveAdminAssignment(courseId: string, studentIds: string[]): Promise<AdminAssignmentRecord | null> {
+  const token = getStoredAuth()?.session_token ?? null;
+  if (!token) return null;
   const response = await request<AdminAssignmentRecord>(`/api/v1/admin/assignments/${courseId}`, {
     method: 'PUT',
     body: JSON.stringify({
       student_ids: studentIds,
     }),
-  });
-  return unwrap(response, () => fallbackRecord(courseId));
+  }, token);
+  return response?.success && response.data ? response.data : null;
 }
